+++ cmd/auth.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"errors"
 	"fmt"
 	"os"
 	"syscall"
@@ -9,6 +10,55 @@ import (
 	"github.com/spf13/cobra"
 	"golang.org/x/term"
 )
+
+var apiKeyPath string = "/.lsw"
+
+func Login() {
+
+	if !isFileExists(apiKeyPath) {
+		writeFile(apiKeyPath, "")
+	}
+
+	LSW.InitLeasewebClient(readfile(apiKeyPath))
+}
+
+func Logout() {
+	writeFile(apiKeyPath, "")
+}
+
+func getHomeDir() string {
+	dirname, err := os.UserHomeDir()
+	if err != nil {
+		fmt.Println("Something went wrong! Home directory not found!")
+		os.Exit(0)
+	}
+	return dirname
+}
+
+func writeFile(path, content string) {
+	if os.WriteFile(getHomeDir()+path, []byte(content), 0644) != nil {
+		fmt.Println("Something went wrong1")
+		os.Exit(0)
+	}
+}
+
+func readfile(path string) string {
+	apiKey, err := os.ReadFile(getHomeDir() + path)
+	if err != nil {
+		fmt.Println("Something went wrong2")
+		os.Exit(0)
+	}
+	return string(apiKey)
+}
+
+func isFileExists(path string) bool {
+	_, err := os.Stat(getHomeDir() + path)
+
+	if err == nil {
+		return true
+	}
+	return !errors.Is(err, os.ErrNotExist)
+}
 
 func init() {
 	rootCmd.AddCommand(loginCmd)
@@ -27,7 +77,7 @@ var loginCmd = &cobra.Command{
 			fmt.Println("Something went wrong")
 			os.Exit(0)
 		}
-		WriteApiKey(string(apiKey))
+		writeFile(apiKeyPath, string(apiKey))
 		Login()
 		_, err = LSW.CustomerAccountApi{}.Get()
 		if err != nil {
